Server_BE/src/app/migrations/app/3_20250706013454_update.py
from tortoise import BaseDBAsyncClient


async def upgrade(db: BaseDBAsyncClient) -> str:
    return """
        ALTER TABLE "chunk_storage" ADD "duration" DOUBLE PRECISION;"""


async def downgrade(db: BaseDBAsyncClient) -> str:
    return """
        ALTER TABLE "chunk_storage" DROP COLUMN "duration";"""
